\name{.DollarNames.GlobalOptionsFun}
\alias{.DollarNames.GlobalOptionsFun}
\alias{Dot.DollarNames.GlobalOptionsFun}
\title{
The .DollarNames method for the GlobalOptionsFun class
}
\description{
The .DollarNames method for the GlobalOptionsFun class
}
\usage{
\method{.DollarNames}{GlobalOptionsFun}(x, pattern = "")
}
\arguments{

  \item{x}{the object returned by \code{\link{set_opt}} or \code{\link{setGlobalOptions}}.}
  \item{pattern}{pattern, please ignore it.}

}
\details{
This makes the option object looks like a list that it allows
option name completion after \code{$}.
}
\author{
z.gu@dkfz.de
}
\examples{
# There is no example
NULL

}
