% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{.v}
\alias{.v}
\alias{$.InternalOptionValue}
\title{Get other option values}
\usage{
.v(opt_name, name_is_character = NA)

\method{$}{InternalOptionValue}(x, nm)
}
\arguments{
\item{opt_name}{The option name.}

\item{name_is_character}{Please ignore, only used internally.}

\item{x}{should always be written as `.v`.}

\item{nm}{The option name.}
}
\description{
Get other option values
}
\details{
When setting one option, the value can be dependent on other option names.
The current value of other option can be accessed by `.v(nm)` or `.v$nm`.
}
\examples{
opt = set_opt(a = 1, b = function() .v$a*2)
opt$b
opt(a = 2); opt$b
opt(a = 4); opt$b
reset_opt(opt); opt$b
}
