% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGlobalOptions.R
\name{reset_opt}
\alias{reset_opt}
\alias{add_opt}
\title{Helper functions}
\usage{
reset_opt(opt)

add_opt(opt, ...)
}
\arguments{
\item{opt}{The option object returned by [`set_opt()`] or [`setGlobalOptions()`].}

\item{...}{New options.}
}
\description{
Helper functions
}
\details{
`reset_opt()` is identical to `opt(RESET = TRUE)`.

`add_opt()` is identical to `opt(..., ADD = TRUE)`.
}
\examples{
opt = set_opt(a = 1, b = 2)
opt$a = 100; opt$b = 200
opt
reset_opt(opt)
opt
opt = set_opt(a = 1)
add_opt(opt, b = 2)
opt
}
