% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGlobalOptions.R
\name{print.GlobalOptionsFun}
\alias{print.GlobalOptionsFun}
\title{Print options}
\usage{
\method{print}{GlobalOptionsFun}(x, ...)
}
\arguments{
\item{x}{The option object returned by [`set_opt()`] or [`setGlobalOptions()`].}

\item{...}{Other arguments.}
}
\description{
Print options
}
\examples{
opt = set_opt(a = 1, b = "b")
opt
}
