\name{print.GlobalOptionsFun}
\alias{print.GlobalOptionsFun}
\title{
Print the GlobalOptionsFun object
}
\description{
Print the GlobalOptionsFun object
}
\usage{
\method{print}{GlobalOptionsFun}(x, ...)
}
\arguments{

  \item{x}{the option object returned by \code{\link{set_opt}} or \code{\link{setGlobalOptions}}.}
  \item{...}{other arguments}

}
\author{
z.gu@dkfz.de
}
\examples{
# There is no example
NULL
}
