\name{$.GlobalOptionsFun}
\alias{$.GlobalOptionsFun}
\title{
Get option value by dollar symbol
}
\description{
Get option value by dollar symbol
}
\usage{
\method{$}{GlobalOptionsFun}(x, nm)
}
\arguments{

  \item{x}{the function returned by \code{\link{setGlobalOptions}}}
  \item{nm}{a single option name}

}
