\name{$<-.GlobalOptionsFun}
\alias{$<-.GlobalOptionsFun}
\title{
Set option value by dollar symbol
}
\description{
Set option value by dollar symbol
}
\usage{
\method{$}{GlobalOptionsFun}(x, nm) <- value
}
\arguments{

  \item{x}{the function returned by \code{\link{setGlobalOptions}}}
  \item{nm}{a single option name}
  \item{value}{the value which is assigned to the option}

}