\name{set_opt}
\alias{set_opt}
\title{
Produce a function which can get or set global options
}
\description{
Produce a function which can get or set global options
}
\usage{
set_opt(...)
}
\arguments{

  \item{...}{all go to \code{\link{setGlobalOptions}}}

}
\details{
This is just a short name for \code{\link{setGlobalOptions}}.
}
\author{
z.gu@dkfz.de
}
\examples{
# There is no example
NULL
}
